test(agents): split embedded run queue coverage
